Rename old reconstruction module
